Convert awards enpoint to new syntax
